requirements.txt
numpy 
streamlit 
pandas
plotly.express 
sklearn   
